New types & NftFormCard refactor
